+++ src/components/TableCorner/TableCorner.tsx
@@ -1,35 +1,60 @@
-import { CSSProperties, MouseEvent, ReactNode, useCallback, useRef } from 'react'
+import { CSSProperties, ChangeEvent, KeyboardEvent, ReactNode, useCallback, useRef } from 'react'
 import { useCellNavigation } from '../../hooks/useCellsNavigation'
 
 interface Props {
   checked?: boolean
   children?: ReactNode
-  onClick?: (event: MouseEvent) => void
-  showCheckBox?: boolean
+  onCheckboxPress?: () => void
   style?: CSSProperties
   ariaColIndex: number
   ariaRowIndex: number
 }
 
-export default function TableCorner({ children, checked, onClick, showCheckBox, style, ariaColIndex, ariaRowIndex }: Props) {
+export default function TableCorner({ children, checked, onCheckboxPress, style, ariaColIndex, ariaRowIndex }: Props) {
   const ref = useRef<HTMLTableCellElement>(null)
   const { tabIndex, navigateToCell } = useCellNavigation({ ref, ariaColIndex, ariaRowIndex })
-  const handleClick = useCallback((event: MouseEvent) => {
+  const handleClick = useCallback(() => {
     navigateToCell()
-    onClick?.(event)
-  }, [onClick, navigateToCell])
+    onCheckboxPress?.()
+  }, [onCheckboxPress, navigateToCell])
+  const handleKeyDown = useCallback((event: KeyboardEvent) => {
+    if (event.key === 'Enter' || event.key === ' ') {
+      event.preventDefault()
+      // TODO: let the event propagate?
+      event.stopPropagation()
+      onCheckboxPress?.()
+    }
+  }, [onCheckboxPress])
+  const showCheckBox = checked !== undefined
+  const disabledCheckbox = onCheckboxPress === undefined
+  const onChange = useCallback((e: ChangeEvent) => {e.preventDefault()}, [])
 
   return (
     <td
       ref={ref}
-      aria-disabled={!showCheckBox}
       style={style}
       onClick={handleClick}
+      onKeyDown={handleKeyDown}
+      aria-checked={checked}
+      aria-rowindex={ariaRowIndex}
       aria-colindex={ariaColIndex}
+      aria-disabled={disabledCheckbox}
       tabIndex={tabIndex}
     >
-      <span>{children}</span>
-      { showCheckBox && <input type='checkbox' checked={checked} readOnly /> }
+      {
+        showCheckBox ?
+          <input
+            type='checkbox'
+            onChange={onChange}
+            readOnly={disabledCheckbox}
+            disabled={disabledCheckbox}
+            checked={checked}
+            role="presentation"
+            tabIndex={-1}
+          />
+          :
+          <span>{children}</span>
+      }
     </td>
   )
 }